+++ app/src/main/java/com/wei/vsclock/VsclockApplication.kt
@@ -1,0 +1,17 @@
+package com.wei.vsclock
+
+import android.app.Application
+import com.wei.vsclock.core.common.BuildConfig
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class VsclockApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
