+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
-    <string name="app_name">vsclock</string>
+    <string name="app_name">VsClock</string>
+    <string name="not_connected">⚠️ You aren’t connected to the internet</string>
+
 </resources>
