+++ core/designsystem/src/main/java/com/wei/vsclock/core/designsystem/theme/Background.kt
@@ -1,0 +1,20 @@
+package com.wei.vsclock.core.designsystem.theme
+
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.staticCompositionLocalOf
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
+
+/**
+ * A class to model background color and tonal elevation values for VsClock.
+ */
+@Immutable
+data class BackgroundTheme(
+    val color: Color = Color.Unspecified,
+    val tonalElevation: Dp = Dp.Unspecified,
+)
+
+/**
+ * A composition local for [BackgroundTheme].
+ */
+val LocalBackgroundTheme = staticCompositionLocalOf { BackgroundTheme() }
