+++ core/designsystem/src/main/java/com/wei/vsclock/core/designsystem/theme/Color.kt
@@ -1,0 +1,67 @@
+package com.wei.vsclock.core.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+val md_theme_light_primary = Color(0xFF006A66)
+val md_theme_light_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFF29FCF2)
+val md_theme_light_onPrimaryContainer = Color(0xFF00201E)
+val md_theme_light_secondary = Color(0xFF4A6361)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFCCE8E5)
+val md_theme_light_onSecondaryContainer = Color(0xFF051F1E)
+val md_theme_light_tertiary = Color(0xFF49607B)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFD0E4FF)
+val md_theme_light_onTertiaryContainer = Color(0xFF011D34)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFAFDFB)
+val md_theme_light_onBackground = Color(0xFF191C1C)
+val md_theme_light_surface = Color(0xFFFAFDFB)
+val md_theme_light_onSurface = Color(0xFF191C1C)
+val md_theme_light_surfaceVariant = Color(0xFFDAE5E3)
+val md_theme_light_onSurfaceVariant = Color(0xFF3F4948)
+val md_theme_light_outline = Color(0xFF6F7978)
+val md_theme_light_inverseOnSurface = Color(0xFFEFF1F0)
+val md_theme_light_inverseSurface = Color(0xFF2D3131)
+val md_theme_light_inversePrimary = Color(0xFF00DDD5)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFF006A66)
+val md_theme_light_outlineVariant = Color(0xFFBEC9C7)
+val md_theme_light_scrim = Color(0xFF000000)
+
+val md_theme_dark_primary = Color(0xFF00DDD5)
+val md_theme_dark_onPrimary = Color(0xFF003735)
+val md_theme_dark_primaryContainer = Color(0xFF00504D)
+val md_theme_dark_onPrimaryContainer = Color(0xFF29FCF2)
+val md_theme_dark_secondary = Color(0xFFB0CCC9)
+val md_theme_dark_onSecondary = Color(0xFF1B3533)
+val md_theme_dark_secondaryContainer = Color(0xFF324B49)
+val md_theme_dark_onSecondaryContainer = Color(0xFFCCE8E5)
+val md_theme_dark_tertiary = Color(0xFFB0C9E7)
+val md_theme_dark_onTertiary = Color(0xFF19324A)
+val md_theme_dark_tertiaryContainer = Color(0xFF314962)
+val md_theme_dark_onTertiaryContainer = Color(0xFFD0E4FF)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF191C1C)
+val md_theme_dark_onBackground = Color(0xFFE0E3E2)
+val md_theme_dark_surface = Color(0xFF191C1C)
+val md_theme_dark_onSurface = Color(0xFFE0E3E2)
+val md_theme_dark_surfaceVariant = Color(0xFF3F4948)
+val md_theme_dark_onSurfaceVariant = Color(0xFFBEC9C7)
+val md_theme_dark_outline = Color(0xFF889391)
+val md_theme_dark_inverseOnSurface = Color(0xFF191C1C)
+val md_theme_dark_inverseSurface = Color(0xFFE0E3E2)
+val md_theme_dark_inversePrimary = Color(0xFF006A66)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFF00DDD5)
+val md_theme_dark_outlineVariant = Color(0xFF3F4948)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+val seed = Color(0xFF0FF4EB)
