+++ core/designsystem/src/main/java/com/wei/vsclock/core/designsystem/theme/Spacing.kt
@@ -1,0 +1,7 @@
+package com.wei.vsclock.core.designsystem.theme
+
+const val SPACING_EXTRA_SMALL = 4
+const val SPACING_SMALL = 8
+const val SPACING_MEDIUM = 12
+const val SPACING_LARGE = 16
+const val SPACING_EXTRA_LARGE = 24
