+++ core/designsystem/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="core_designsystem_close">CLOSE</string>
+    <string name="core_designsystem_functionality_not_available">Functionality not available 🙈</string>
+</resources>
